pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lexwiki"
version = "0.1.0"
description = "LLM-powered legal knowledge base compiler. Ingest legal documents, compile structured Obsidian wikis, query with natural language, lint for consistency."
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.10"
authors = [
    { name = "AnyLegal", email = "hello@anylegal.ai" },
]
keywords = ["legal", "knowledge-base", "llm", "obsidian", "wiki", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.27.0",
    "pymupdf4llm>=0.0.25",
    "mammoth>=1.6.0",
    "trafilatura>=1.6.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.21",
    "pyyaml>=6.0",
    "rich>=13.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]

[project.urls]
Homepage = "https://anylegal.ai"
Repository = "https://github.com/wouldbe12/lexwiki"

[project.scripts]
lexwiki = "lexwiki.cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["src/lexwiki"]

[tool.ruff]
target-version = "py310"
line-length = 100
